ffmpeg/qsv: move VP9-10 Encoder Test classes to lib
